requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
streamlit==1.35.0
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.7.4